Move QAppBarMenu to own file
